Fix ONSAM-1953

+++ Publications/GPU-Opt-Guide/conditionals/convolution-global-conditionals-min-max.cpp
@@ -3,14 +3,14 @@
 //
 // SPDX-License-Identifier: MIT
 // =============================================================
-#include <CL/sycl.hpp>
 #include <iostream>
 #include <random>
+#include <sycl/sycl.hpp>
 #include <vector>
 
 int main() {
-  constexpr size_t N = 8192 * 8192;
-  constexpr size_t M = 257;
+  constexpr int N = 8192 * 8192;
+  constexpr int M = 257;
 
   std::vector<int> input(N);
   std::vector<int> output(N);
